extract rescue function into rescue.py
