+++ savepoints/services/snapshot.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import bpy
 
-from .storage import get_history_dir, load_manifest
+from .storage import get_history_dir, load_manifest, SNAPSHOT_FILENAME, THUMBNAIL_FILENAME
 from .thumbnail import capture_thumbnail
 from .versioning import add_version_to_manifest
 from ..ui_utils import sync_history_to_props
@@ -24,15 +24,13 @@ def create_snapshot(context, version_id, note, skip_thumbnail=False):
     obj_count = len(bpy.data.objects)
 
     # Thumbnail
-    thumb_filename = "thumbnail.png"
+    thumb_filename = THUMBNAIL_FILENAME
     thumb_path = version_dir / thumb_filename
     if not skip_thumbnail:
         capture_thumbnail(context, str(thumb_path))
 
     # Save Snapshot
-    blend_filename = "snapshot.blend_snapshot"
-    snapshot_path = version_dir / blend_filename
-
+    snapshot_path = version_dir / SNAPSHOT_FILENAME
     bpy.ops.wm.save_as_mainfile(copy=True, filepath=str(snapshot_path))
 
     # Capture file size
@@ -46,7 +44,7 @@ def create_snapshot(context, version_id, note, skip_thumbnail=False):
         version_id,
         note,
         str(Path(folder_name) / thumb_filename),
-        str(Path(folder_name) / blend_filename),
+        str(Path(folder_name) / SNAPSHOT_FILENAME),
         object_count=obj_count,
         file_size=file_size
     )
