+++ tests/blender/test_quick_save_edit_note.py
@@ -10,7 +10,6 @@ ROOT = Path(__file__).resolve().parents[2]
 sys.path.append(str(ROOT))
 
 import savepoints  # noqa: E402
-from savepoints import core
 from savepoints.services.storage import load_manifest
 
 def setup_test_env():
